connect position to stock

+++ server/reactivestocks/api/views.py
@@ -1,3 +1,4 @@
+from turtle import pos, st, up
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
@@ -41,7 +42,7 @@ def update(stock):
 def get_followed_stocks(request):
     if not Stock.objects.exists():
         return Response(status=status.HTTP_204_NO_CONTENT)
-    stocks = Stock.objects.all()
+    stocks = Stock.objects.filter(followed=True)
     for stock in stocks:
         update(stock)
     serializedData = StockSerializer(Stock.objects.all(), many=True).data
@@ -58,13 +59,22 @@ def get_suggestions(request, symbol):
 @api_view(['POST'])
 def add_stock(request):
     symbol = request.data['symbol']
-    data = requests.get(
-        f'https://financialmodelingprep.com/api/v3/quote/{symbol}?apikey={APIKEY}').json()
-    serializer = StockSerializer(data=data[0])
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if Stock.objects.filter(symbol=symbol).exists():
+        if Stock.objects.get(symbol=symbol).followed:
+            return Response(status=status.HTTP_409_CONFLICT)
+        stock = Stock.objects.get(symbol=symbol)
+        stock.followed = True
+        stock.save()
+        return Response(StockSerializer(stock).data, status=status.HTTP_201_CREATED)
+    else:
+        data = requests.get(
+            f'https://financialmodelingprep.com/api/v3/quote/{symbol}?apikey={APIKEY}').json()
+        data[0]['followed'] = True
+        serializer = StockSerializer(data=data[0])
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['DELETE'])
@@ -73,31 +83,61 @@ def remove_stock(request, pk):
         stock = Stock.objects.get(pk=pk)
     except Stock.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-    stock.delete()
+    if not stock.owned:
+        stock.delete()
+    else:
+        stock.followed = False
+        stock.save()
     return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['GET'])
 def get_positions(request):
-    return Response(PositionSerializer(Position.objects.all(), many=True).data)
+    if not Position.objects.exists():
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    positions = Position.objects.all()
+    for position in positions:
+        stock = position.Stock
+        update(stock)
+    serializedData = PositionSerializer(positions, many=True).data
+    return Response(serializedData)
 
 
 @api_view(['POST'])
 def open_position(request):
     data = request.data['positionData']
-    position = Position.objects.filter(
-        symbol=data['symbol']).first()
-    if position is not None:
-        position.average_price = (position.average_price * position.quantity +
-                                  float(data['average_price']) * float(data['quantity']))/(position.quantity + float(data['quantity']))
-        position.quantity += float(data['quantity'])
-        position.save()
-        return Response(PositionSerializer(position).data, status=status.HTTP_201_CREATED)
-    serializer = PositionSerializer(data=data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    stock = Stock.objects.filter(symbol=data['symbol']).first()
+    del data['symbol']
+    if stock is None:
+        stock_data = requests.get(
+            f'https://financialmodelingprep.com/api/v3/quote/{symbol}?apikey={APIKEY}').json()
+        stock_data[0]['owned'] = True
+        serializer = StockSerializer(data=data[0])
+        if serializer.is_valid():
+            serializer.save()
+            stock = Stock.objects.filter(symbol=data['symbol']).first()
+            data['Stock'] = stock
+            serializer = PositionSerializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+        position = Position.objects.filter(Stock=stock).first()
+        if position is not None:
+            position.average_price = (position.average_price * position.quantity +
+                                      float(data['average_price']) * float(data['quantity']))/(position.quantity + float(data['quantity']))
+            position.quantity += float(data['quantity'])
+            position.save()
+            return Response(PositionSerializer(position).data, status=status.HTTP_201_CREATED)
+        serializer = PositionSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            stock.owned = True
+            stock.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['DELETE'])
@@ -106,5 +146,7 @@ def close_position(request, pk):
         position = Position.objects.get(pk=pk)
     except Position.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+    position.Stock.owned = False
+    position.Stock.save()
     position.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
